Add optional ORCID provenance column (#31)
Closes #30 

This PR does the following:

- [x] Add the ability to add an **optional** ORCID column (labeled
"contributor") to each curated TSV
- [x] Adds two example ORCID annotations
- [x] Updates the output to incorporate ORCID provenance in SSSOM and
ontology output
- [x] Update data integrity tests (will do this if the PR gets positive
feedback and the questions below are answered)
 
It has the minor drawback of creating a big diff, but it's mostly
localized to generated artifacts.

Questions before merging this, if everyone thinks it's a good idea:

1. Should we enforce that the `contributor` column is always present,
even if it's mostly empty?
2. Should we curate a list of ORCID->Name mappings to make these outputs
(namely, the ontology) nicer to read in Protege? It's also possible to
quickly interact with the ORCID REST API to look up names as an
alternative.